Added exceptions for Producto

+++ src/main/java/shop/services/ProductoService.java
@@ -1,10 +1,12 @@
 package shop.services;
 
+import org.springframework.web.servlet.HandlerExceptionResolver;
 import shop.model.Producto;
 import shop.repos.ProductRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import javax.validation.constraints.Null;
 import java.util.List;
 
 @Service
@@ -14,18 +16,25 @@ public class ProductoService {
     private ProductRepo products;
 
     public Producto postProduct(Producto p){
+        if(p.getItem().isEmpty()){
+            throw new NullPointerException();
+        }
         return products.save(p);
     }
 
     public Producto deleteProducto(int id){
-        Producto p = products.findById(id).get();
-        products.deleteById(id);
-        return p;
+        if(products.existsById(id)){
+            Producto p = products.findById(id).get();
+            products.deleteById(id);
+            return p;
+        } else throw new IllegalArgumentException();
     }
 
     public Producto getProductById(int id){
-        Producto p = products.findById(id).get();
-        return p;
+        if(products.existsById(id)){
+            Producto p = products.findById(id).get();
+            return p;
+        } else throw new IllegalArgumentException();
     }
 
     public List<Producto> findAll(){
@@ -33,7 +42,9 @@ public class ProductoService {
     }
 
     public Producto putProducto(int id, Producto p){
-        p.setId(id);
-        return products.save(p);
+        if(products.existsById(id)){
+            p.setId(id);
+            return products.save(p);
+        } else throw new IllegalArgumentException();
     }
 }
